requirements.txt
anthropic>=0.40.0
openai-whisper>=20231117
pyaudio>=0.2.14
pvporcupine>=3.0.0
gTTS>=2.5.0
RPi.GPIO>=0.7.1
gpiozero>=2.0
